# LIS_backend/LIS_API/LIS_API/DTOs/ProductDTO.cs
namespace LIS_API.DTOs
{
    public class ProductDTO
    {
        public int id { get; set; }
        public int category { get; set; }
        public string name { get; set; }
        public string code { get; set; }
        public string description { get; set; }
        public string created { get; set; }
        public string updated { get; set; }
    }
}