# LIS_backend/LIS_API/LIS_API/Models/Invoice.cs
using System.ComponentModel.DataAnnotations;

namespace LIS_API.Models
{
    public class Invoice
    {
        [Key]
        public int Id { get; set; }
        public string date { get; set; }
        public string status { get; set; }
        public User user { get; set; }
        public string created {  get; set; }
        public string updated { get; set; }
    }
}