# LIS_backend/LIS_API/LIS_API/Models/User.cs
using System.ComponentModel.DataAnnotations;

namespace LIS_API.Models
{
    public class User
    {
        [Key]
        public int id { get; set; }
        public string username { get; set; }
        public string password { get; set; }
        public string role { get; set; }
    }
}